requirements.txt
firebase-admin
google-cloud-firestore
bcrypt
python-dotenv
pillow
opencv-python
ultralytics
tk
